# openapi/openapi.yaml
components:
  schemas:
...
    handlers.CreateMemberAppRequest:
      properties:
        app_name:
          description: AppName 是新实例名称，创建时必填。
          type: string
        app_prompt:
          description: AppPrompt 是新实例提示词，可为空。
          type: string
        channel_type:
          description: ChannelType 是初始化渠道绑定的渠道标识。
          type: string
        model_id:
          description: ModelID 是新实例使用的模型，必须属于组织 enabled_models。
          type: string
        persona_mode:
          description: PersonaMode 控制新实例是否继承组织人设或使用独立人设。
          type: string
        runtime_node_id:
          description: NodeID 是指定 runtime 节点；为空时 service 自动选择可用节点。
          type: string
      required:
      - app_name
      - model_id
      type: object
    handlers.CreateMemberRequest:
      properties:
        display_name:
          description: DisplayName 是前端显示名，创建与更新时都不能为空。
          type: string
        password:
          description: Password 是初始密码，service 层立即 hash 后写库。
          type: string
        role:
          description: Role 允许 org_admin / org_member；空值由 service 层补默认 org_member。
          type: string
        username:
          description: Username 是组织内登录账号名，创建后不可通过成员更新接口修改。
          type: string
      required:
      - display_name
      - password
      - username
      type: object
...
    handlers.ErrorResponse:
      properties:
        code:
          description: Code 是稳定的接口契约标识，一旦发布只增不改。
          example: APP_NOT_FOUND
          type: string
        message:
          description: Message 是面向前端展示的安全错误文案。
          example: 应用不存在
          type: string
      type: object
...
    handlers.OnboardMemberRequest:
      properties:
        app_name:
          description: AppName 是随成员初始化的默认应用名称。
          type: string
        app_prompt:
          description: AppPrompt 是默认应用的初始提示词，可为空。
          type: string
        channel_type:
          description: ChannelType 是初始化渠道绑定的渠道标识。
          type: string
        display_name:
          description: DisplayName 是新成员展示名。
          type: string
        model_id:
          description: ModelID 是新实例使用的模型，必须属于组织 enabled_models。
          type: string
        password:
          description: Password 是新成员初始密码。
          type: string
        persona_mode:
          description: PersonaMode 控制新应用是否继承组织人设或使用独立人设。
          type: string
        role:
          description: Role 为空时默认为 org_member；不允许创建 platform_admin。
          type: string
        runtime_node_id:
          description: NodeID 是指定 runtime 节点；为空时 service 自动选择可用节点。
          type: string
        username:
          description: Username 是新成员账号名，与普通创建成员保持同一约束。
          type: string
      required:
      - app_name
      - display_name
      - model_id
      - password
      - username
      type: object
...
    handlers.ResetPasswordRequest:
      properties:
        password:
          description: Password 是新密码，service 层只保存 hash。
          type: string
      required:
      - password
      type: object
...
    handlers.UpdateMemberRequest:
      properties:
        display_name:
          description: DisplayName 是成员展示名，更新接口要求显式传入非空值。
          type: string
        role:
          description: Role 为空表示保持原角色；非空时需要管理员权限并限制在组织角色内。
          type: string
      required:
      - display_name
      type: object
...
    service.CreateAppForMemberResult:
      properties:
        app:
          $ref: '#/components/schemas/service.AppResult'
        job_id:
          type: string
      type: object
...
    service.MemberResult:
      properties:
        display_name:
          description: DisplayName 是前端展示名。
          type: string
        id:
          description: ID 是成员用户 UUID。
          type: string
        org_id:
          description: OrgID 是成员所属组织 UUID；platform_admin 可能为空。
          type: string
        role:
          description: Role 是成员角色，限定为 org_admin 或 org_member。
          type: string
        status:
          description: Status 是成员状态；disabled 会阻止登录并设置 users.deleted_at。
          type: string
        username:
          description: Username 是登录账号名。
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 登录后获得的 JWT access token，前缀 "Bearer "。
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 基于 Hermes Agent runtime 的多组织管理后台 API。
  title: Agent Runtime Manager API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /members/{userId}:
    delete:
      description: 软删指定成员，并联动触发应用回收任务
      parameters:
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: 删除成功，无响应体
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 删除成员
      tags:
      - members
    get:
      description: 按 userId 获取单个成员信息；org_member 只能查询自身
      parameters:
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.MemberResult'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 成员详情
      tags:
      - members
    patch:
      description: 更新成员显示名或角色；org_admin 仅可更新自己组织的成员
      parameters:
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.UpdateMemberRequest'
                description: 更新成员请求
                summary: body
        description: 更新成员请求
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.MemberResult'
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 更新成员
      tags:
      - members
  /members/{userId}/disable:
    post:
      description: 将成员状态设为 disabled，成员登录时将被拒绝
      parameters:
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.MemberResult'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 禁用成员
      tags:
      - members
  /members/{userId}/enable:
    post:
      description: 将成员状态从 disabled 恢复为 active
      parameters:
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.MemberResult'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 启用成员
      tags:
      - members
  /members/{userId}/password:
    post:
      description: 组织管理员强制重置本组织成员的密码
      parameters:
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.ResetPasswordRequest'
                description: 重置密码请求
                summary: body
        description: 重置密码请求
        required: true
      responses:
        "204":
          description: 密码重置成功，无响应体
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 重置成员密码
      tags:
      - members
...
  /organizations/{orgId}/members:
    get:
      description: 按组织 ID 分页列出成员；org_member 只能看到自己
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      - description: 每页条数（默认不限）
        in: query
        name: limit
        schema:
          type: integer
      - description: 分页偏移（默认 0）
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/service.MemberResult'
                  type: array
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 组织成员列表
      tags:
      - members
    post:
      description: 组织管理员在本组织下创建新成员
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.CreateMemberRequest'
                description: 创建成员请求
                summary: body
        description: 创建成员请求
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.MemberResult'
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 创建组织成员
      tags:
      - members
  /organizations/{orgId}/members/{userId}/apps:
    post:
      description: 平台管理员或本组织管理员为已有成员创建新的应用实例；目标成员必须没有未删除实例
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      - description: 成员用户 ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.CreateMemberAppRequest'
                description: 创建实例请求
                summary: body
        description: 创建实例请求
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.CreateAppForMemberResult'
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Service Unavailable
      security:
      - BearerAuth: []
      summary: 为已有成员创建实例
      tags:
      - members
  /organizations/{orgId}/members/onboard:
    post:
      description: 在单个事务内创建成员并联动初始化应用、渠道绑定和 newapi 凭证；onboarding service 未配置时返回 503
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.OnboardMemberRequest'
                description: onboarding 请求
                summary: body
        description: onboarding 请求
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.OnboardMemberResult'
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Service Unavailable
      security:
      - BearerAuth: []
      summary: 成员 onboarding（事务创建）
      tags:
      - members
...
servers:
- url: /api/v1